Medical_chatbot/requirements.txt
sentence-transformers==2.7.0
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-classic
langchain-huggingface
langchain-nvidia-ai-endpoints
